# internal/docs/swagger.yaml
definitions:
...
  api.CreateTextTrackResponse:
    properties:
      text_track:
        $ref: '#/definitions/api.TextTrackResponse'
    required:
    - text_track
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    required:
    - error
    type: object
...
  api.ListTextTracksResponse:
    properties:
      text_tracks:
        items:
          $ref: '#/definitions/api.TextTrackResponse'
        type: array
    required:
    - text_tracks
    type: object
...
  api.TextTrackResponse:
    properties:
      asset_id:
        type: string
      asset_version_id:
        type: string
      content:
        type: string
      content_truncated:
        type: boolean
      created_at:
        type: string
      download_url:
        type: string
      error:
        type: string
      has_file:
        type: boolean
      id:
        type: string
      lang:
        type: string
      meta:
        additionalProperties: {}
        type: object
      source:
        type: string
      status:
        type: string
      updated_at:
        type: string
    required:
    - asset_id
    - content
    - content_truncated
    - created_at
    - has_file
    - id
    - meta
    - source
    - status
    - updated_at
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/assets/{id}/text-tracks:
    get:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ListTextTracksResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List text tracks for an asset
      tags:
      - Text Tracks
    post:
      consumes:
      - application/json
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CreateTextTrackResponse'
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.CreateTextTrackResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a text track for an asset
      tags:
      - Text Tracks
  /api/v1/assets/{id}/text-tracks/{tid}:
    delete:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Text Track ID
        in: path
        name: tid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Text track not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a text track
      tags:
      - Text Tracks
    get:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Text Track ID
        in: path
        name: tid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TextTrackResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Text track not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a text track
      tags:
      - Text Tracks
...
swagger: "2.0"